program_stock/requirements.txt
# 量化交易系统依赖
backtrader>=1.9.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
pymongo>=4.0.0
pyyaml>=6.0
talib>=0.4.0  # 可选，技术指标库（需先安装TA-Lib系统库）
